Grammars for generating ironic movie titles
Two Tracery grammars for generating silly movie titles. The ironic generator pairs a movie genre and quality to a title that clearly evinces the opposite. The sequel generator works on the premise that sequels betray the promise of the original and become increasingly unmoored from the originating principle. The sequel titles pair a serious title (apt) with an inept chapter name to show the steady decline in Hollywood sanity.