Tracery grammar for craft-beers and gastropubs
The craft-beer grammar pokes fun at the craft beer movement, not at the beer itself but at the names of the new beers and the establishments that sell them.  The names are generated by turning the vivid descriptions found in ironic similes into product names using simple capitalisation and formatting. The pub names follow the standard English "The X and Y" template discussed in our book on Twitterbots (chapter 2). Again these are generated by transforming the funny phrases that can be mined from ironic similes. Here's a good example:

I've just been down my local Gastropub, The Porkchop And Synagogue, for a bucket of pig swill.