Conspiracy generators
Tracery grammars to generate accusations and cover-ups and to frame them as conspiracy theories. Two grammars can be found here:  accusation grammar.txt, which generates the base accusations and provide hooks (initially empty) through which the accusation can be framed as a conspiracy theory; the grammar   conspiracy 101.txt   instantiates these hooks to frame the accusations as right-wing conspiracies.   The hook rules begin with the rule "pre" and provide expansions for the victims and for the string-pullers.